+++ packages/lumx-react/src/components/slideshow/Slides.tsx
@@ -3,11 +3,41 @@ import React, { CSSProperties, forwardRef } from 'react';
 import classNames from 'classnames';
 
 import { FULL_WIDTH_PERCENT } from '@lumx/react/components/slideshow/constants';
-import { Comp, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
+import { Comp, GenericProps, getRootClassName, handleBasicClasses } from '@lumx/react/utils';
 import { mergeRefs } from '@lumx/react/utils/mergeRefs';
+import { Theme } from '@lumx/react';
 
-export interface SlidesProps {
-
+export interface SlidesProps extends GenericProps {
+    /** current slide active */
+    activeIndex?: number;
+    /** slides id to be added to the wrapper */
+    id?: string;
+    /** callback to set a reference to the slideshow */
+    setSlideshow: (slideshow: HTMLDivElement | undefined) => void;
+    /** custom classname */
+    className?: string;
+    /** custom theme */
+    theme?: Theme;
+    /** Whether the automatic rotation of the slideshow is enabled or not. */
+    autoPlay?: boolean;
+    /** Whether the image has to fill its container height or not. */
+    fillHeight?: boolean;
+    /** Number of slides to group together. */
+    groupBy?: number;
+    /** Interval between each slide when automatic rotation is enabled. */
+    interval?: number;
+    /** whether the slides are currently playing or not */
+    isAutoPlaying?: boolean;
+    /** id to be passed in into the slides */
+    slidesId?: string;
+    /** callback to change whether the slideshow is playing or not */
+    setIsAutoPlaying: (isAutoPlaying: boolean) => void;
+    /** starting visible index */
+    startIndexVisible: number;
+    /** ending visible index */
+    endIndexVisible: number;
+    /** component to be rendered after the slides */
+    afterSlides?: React.ReactNode;
 }
 
 /**
@@ -30,7 +60,7 @@ const CLASSNAME = getRootClassName(COMPONENT_NAME);
 export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
     const {
         activeIndex,
-        slideshowId,
+        id,
         setSlideshow,
         className,
         theme,
@@ -38,12 +68,13 @@ export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
         groupBy,
         isAutoPlaying,
         autoPlay,
-        slideshowSlidesId,
+        slidesId,
         setIsAutoPlaying,
         startIndexVisible,
         endIndexVisible,
         children,
         afterSlides,
+        interval,
         ...forwardedProps
     } = props;
     // Inline style of wrapper element.
@@ -52,7 +83,7 @@ export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
     /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
     return (
         <section
-            id={slideshowId}
+            id={id}
             ref={mergeRefs(ref, setSlideshow)}
             {...forwardedProps}
             className={classNames(className, handleBasicClasses({ prefix: CLASSNAME, theme }), {
@@ -63,7 +94,7 @@ export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
             aria-live={isAutoPlaying ? 'off' : 'polite'}
         >
             <div
-                id={slideshowSlidesId}
+                id={slidesId}
                 className={`${CLASSNAME}__slides`}
                 onMouseEnter={() => setIsAutoPlaying(false)}
                 onMouseLeave={() => setIsAutoPlaying(Boolean(autoPlay))}
@@ -71,13 +102,11 @@ export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
                 <div className={`${CLASSNAME}__wrapper`} style={wrapperStyle}>
                     {React.Children.map(children, (child: React.ReactNode, index: number) => {
                         if (React.isValidElement(child)) {
-                            const isCurrentlyVisible = index >= startIndexVisible && index <= endIndexVisible;
+                            const isCurrentlyVisible = index >= startIndexVisible && index < endIndexVisible;
 
                             return React.cloneElement(child, {
-                                style: !isCurrentlyVisible
-                                    ? { visibility: 'hidden', ...(child.props.style || {}) }
-                                    : child.props.style || {},
-                                'aria-hidden': !isCurrentlyVisible,
+                                isCurrentlyVisible,
+                                interval,
                             });
                         }
 
@@ -90,3 +119,6 @@ export const Slides: Comp<any, HTMLDivElement> = forwardRef((props, ref) => {
         </section>
     );
 });
+
+Slides.displayName = COMPONENT_NAME;
+Slides.className = CLASSNAME;